refactor: Add Keypairs as fixtures
