+++ apps/sdk-e2e/src/integration/transaction-kinetic-sdk-e2e.spec.ts
@@ -1,0 +1,104 @@
+import { MogamiSdk } from '@mogami/sdk'
+import { Keypair } from '@mogami/keypair'
+import { aliceKeypair, bobKeypair, charlieKeypair, daveKeypair } from './fixtures'
+import { Destination } from '@mogami/solana'
+import { AppTransactionStatus } from '@prisma/client'
+
+describe('MogamiSdk (e2e)', () => {
+  let sdk: MogamiSdk
+  const defaultMint = process.env.DEFAULT_MINT_PUBLIC_KEY
+
+  beforeEach(async () => {
+    sdk = await MogamiSdk.setup({ index: 1, endpoint: 'http://localhost:3000', environment: 'devnet' })
+  })
+
+  it('should make a transfer', async () => {
+    const tx = await sdk.makeTransfer({ amount: '43', destination: bobKeypair.publicKey, owner: aliceKeypair })
+    expect(tx).not.toBeNull()
+    expect(tx.mint).toBe(defaultMint)
+    const { signature, errors, amount, source } = tx
+    expect(typeof signature).toBe('string')
+    expect(errors).toEqual([])
+    expect(Number(amount)).toBe(4300000)
+    expect(source).toBe(aliceKeypair.publicKey)
+  })
+
+  it('should make a transfer in batch', async () => {
+    const destinations: Destination[] = [
+      { destination: bobKeypair.publicKey, amount: '51' },
+      { destination: charlieKeypair.publicKey, amount: '72' },
+      { destination: daveKeypair.publicKey, amount: '87' },
+    ]
+
+    const tx = await sdk.makeTransferBatch({ destinations, owner: aliceKeypair })
+    expect(tx).not.toBeNull()
+    expect(tx.mint).toBe(defaultMint)
+    const { signature, errors, amount, source } = tx
+    expect(typeof signature).toBe('string')
+    expect(errors).toEqual([])
+    expect(Number(amount)).toBe(5100000)
+    expect(source).toBe(aliceKeypair.publicKey)
+  }, 60000)
+
+  it('should throw an error when there are less than 1 transactions in a batch', async () => {
+    try {
+      const destinations: Destination[] = []
+      await sdk.makeTransferBatch({ destinations, owner: aliceKeypair })
+    } catch (error) {
+      expect(error.toString()).toContain('Error: At least 1 destination required')
+    }
+  })
+
+  it('should throw an error when there are more than 15 transactions in a batch', async () => {
+    try {
+      const destinations: Destination[] = []
+      const destination = { destination: bobKeypair.publicKey, amount: '15' }
+      for (let i = 0; i <= 15; i++) {
+        destinations.push(destination)
+      }
+      await sdk.makeTransferBatch({ destinations, owner: aliceKeypair })
+    } catch (error) {
+      expect(error.toString()).toContain('Error: Maximum number of destinations exceeded')
+    }
+  })
+
+  it('should fail when one account does not exist in batch transfer', async () => {
+    try {
+      const destinations: Destination[] = []
+      destinations.push({ destination: bobKeypair.publicKey, amount: '15' })
+      const kp = Keypair.random()
+      destinations.push({ destination: kp.publicKey, amount: '12' })
+      await sdk.makeTransferBatch({ destinations, owner: aliceKeypair })
+    } catch (error) {
+      const errorData = error.response.data.error
+      expect(errorData).toContain("type: 'InvalidAccount'")
+      expect(errorData).toContain("instruction: '2")
+    }
+  })
+
+  it('should throw when insufficient funds in a transaction', async () => {
+    const res = await sdk.makeTransfer({
+      amount: '99999999999999',
+      destination: bobKeypair.publicKey,
+      owner: aliceKeypair,
+    })
+    expect(res.signature).toBeNull()
+    expect(res.amount).toBeNull()
+    expect(res.errors.length).toBeGreaterThan(0)
+    expect(res.status).toBe(AppTransactionStatus.Failed)
+    expect(res.errors[0].message).toContain('Error: failed to send transaction') // TODO: Insufficient funds
+  })
+
+  it('should throw when insufficient funds in a batch transaction', async () => {
+    const destinations: Destination[] = []
+    destinations.push({ destination: bobKeypair.publicKey, amount: '15' })
+    const kp = Keypair.random()
+    destinations.push({ destination: kp.publicKey, amount: '99999999999999' })
+    const res = await sdk.makeTransferBatch({ destinations, owner: aliceKeypair })
+    expect(res.amount).toBeNull()
+    expect(res.signature).toBeNull()
+    expect(res.errors.length).toBeGreaterThan(0)
+    expect(res.status).toBe(AppTransactionStatus.Failed)
+    expect(res.errors[0].message).toContain('Error: failed to send transaction') // TODO: Insufficient funds
+  })
+})
