+++ libs/sdk/src/lib/mogami-sdk.ts
@@ -1,6 +1,4 @@
-import { TransactionType } from '@kin-tools/kin-memo'
-import { Keypair } from '@mogami/keypair'
-import { Commitment, Destination, Solana } from '@mogami/solana'
+import { Solana } from '@mogami/solana'
 import { Cluster, clusterApiUrl } from '@solana/web3.js'
 import { AppTransaction } from '../generated'
 import { getSolanaRpcEndpoint } from './helpers'
